Enhance user management: add user deletion endpoint, refactor user data retrieval, and implement trace ID logging in user authentication

+++ main.py
@@ -3,8 +3,13 @@ import asyncio
 
 from app.initial_data import main
 from app.core.db_config import delete_all_data
+from app.core.config import settings
 
 if __name__ == "__main__":
     main()
     uvicorn.run("app.main:app", host="127.0.0.1", port=8000, reload=True)
-    asyncio.run(delete_all_data())
+    if settings.ENVIRONMENT == "local":
+        asyncio.run(delete_all_data())
+
+    if settings.ENVIRONMENT == "test":
+        asyncio.run(delete_all_data())
